+++ src/lib.rs
@@ -43,7 +43,9 @@ pub mod tls13utils;
 
 pub use tls13api::{Client, Server};
 // Debug exports only
-pub use tls13formats::{get_alert_description, get_alert_level, get_hs_type, ContentType};
+pub use tls13formats::{
+    get_alert_description, get_alert_level, handshake_data::get_hs_type, ContentType,
+};
 
 // === Public API that is NOT in hacspec
