+++ src/tls13api.rs
@@ -10,7 +10,11 @@
 use rand::{CryptoRng, RngCore};
 
 use crate::{
-    server::ServerDB, tls13crypto::*, tls13formats::*, tls13handshake::*, tls13record::*,
+    server::ServerDB,
+    tls13crypto::*,
+    tls13formats::{handshake_data::HandshakeType, *},
+    tls13handshake::*,
+    tls13record::*,
     tls13utils::*,
 };
 
@@ -24,7 +28,7 @@ pub enum Client {
         ClientPostServerHello,
         Option<ClientCipherState0>,
         DuplexCipherStateH,
-        HandshakeData,
+        handshake_data::HandshakeData,
     ),
 
     /// The client handshake state after finishing the handshake.
@@ -96,7 +100,7 @@ impl Client {
             Self::Client0(state, cipher_state) => {
                 let sf = get_handshake_record(handshake_bytes)?;
                 let (cipher1, cstate) = client_set_params(&sf, state)?;
-                let buf = HandshakeData::from(Bytes::new());
+                let buf = handshake_data::HandshakeData::from(Bytes::new());
                 Ok((None, Self::ClientH(cstate, cipher_state, cipher1, buf)))
             }
             Self::ClientH(cstate, cipher0, cipher_hs, buf) => {
